requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic>=2.0.0
celery==5.2.7
redis==3.5.3
requests==2.26.0
python-dotenv==1.0.0
pytest==6.2.5
httpx==0.19.0
hupper==1.12.1

